# docs/swagger.yaml
basePath: /
definitions:
  request.Login:
    properties:
      checkCode:
        type: string
      checkCodeKey:
        type: string
      email:
        type: string
      password:
        type: string
    required:
    - checkCode
    - checkCodeKey
    - email
    - password
    type: object
  request.Register:
    properties:
      checkCode:
        type: string
      checkCodeKey:
        type: string
      email:
        type: string
      password:
        type: string
      userName:
        maxLength: 32
        minLength: 5
        type: string
    required:
    - checkCode
    - checkCodeKey
    - email
    - password
    - userName
    type: object
  request.UserInfo:
    properties:
      info:
        type: string
    required:
    - info
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      token:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: msp060308@gmail.com
    name: markus
  description: SeaChat API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  version: "1.0"
paths:
  /api/account/getcaptcha:
    get:
      consumes:
      - application/json
      description: 获取验证码
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取验证码
      tags:
      - Account
  /api/account/getuserinfo:
    post:
      consumes:
      - application/json
      description: 获取用户信息
      parameters:
      - description: 用户信息
        in: body
        name: userinfo
        required: true
        schema:
          $ref: '#/definitions/request.UserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取用户信息
      tags:
      - Account
  /api/account/login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: 用户登录
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户登录
      tags:
      - Account
  /api/account/logout:
    post:
      consumes:
      - application/json
      description: 用户登出
      parameters:
      - description: 用户信息
        in: body
        name: userinfo
        required: true
        schema:
          $ref: '#/definitions/request.UserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户登出
      tags:
      - Account
  /api/account/register:
    post:
      consumes:
      - application/json
      description: 用户注册
      parameters:
      - description: 用户注册信息
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/request.Register'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册
      tags:
      - Account
  /friend/accept:
    post:
      consumes:
      - application/json
      description: 接受好友(同意请求)
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户信息
        in: body
        name: userInfo
        required: true
        schema:
          $ref: '#/definitions/request.UserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 接受好友(同意请求)
      tags:
      - Friend
  /friend/add:
    post:
      consumes:
      - application/json
      description: 添加好友(发送请求)
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户信息
        in: body
        name: userInfo
        required: true
        schema:
          $ref: '#/definitions/request.UserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 添加好友(发送请求)
      tags:
      - Friend
  /health:
    get:
      consumes:
      - application/json
      description: 获取健康状态
      produces:
      - application/json
      responses:
        "200":
          description: 返回结果
          schema: {}
      summary: 获取健康状态
      tags:
      - 服务管理
swagger: "2.0"